refactor. 평가 입력 조립을 EndedChatRoomLoader 로 분리
EndedChatReviewOpener 가 "언제 열지·실패를 어떻게 격리할지"와 "원본 매칭을 타고
입력을 조립할지" 두 층위를 함께 들고 있었다. 리포지토리 4개가 후자에만 쓰이는 것이
그 신호였다. RematchPairCreator 를 뽑을 때와 같은 이유로 협력자를 분리한다.

주입 8개 → 5개, 함수 10개 → 5개.
처리 대상 타입 차이가 드러나도록 openEach·openOne 을 openRooms·openRoom 으로 바꾼다.

Refs #129
