+++ api/src/main/kotlin/com/ditto/api/review/service/EndedChatReviewOpener.kt
@@ -2,29 +2,22 @@ package com.ditto.api.review.service
 
 import com.ditto.api.review.dto.EndedChatRoom
 import com.ditto.domain.chat.entity.ChatRoom
-import com.ditto.domain.chat.entity.ChatRoomType
-import com.ditto.domain.chat.repository.ChatRoomMemberRepository
 import com.ditto.domain.chat.repository.ChatRoomRepository
-import com.ditto.domain.match.repository.GroupMatchRepository
-import com.ditto.domain.match.repository.PersonalMatchRepository
-import com.ditto.domain.quiz.repository.QuizSetRepository
 import com.ditto.domain.review.repository.MemberReviewRepository
 import io.github.oshai.kotlinlogging.KotlinLogging
-import java.time.LocalDate
 import org.springframework.stereotype.Component
 
 /**
  * 끝난 채팅으로 평가를 연다 — 채팅 종료 트랙과 평가 트랙을 잇는 어댑터.
  *
- * 채팅 쪽은 평가를 알지 않는다(`ChatRoomEndService`). 이 어댑터가 종료 결과를 [EndedChatRoom]으로
+ * 채팅 쪽은 평가를 알지 않는다(`ChatRoomEndService`). 이 어댑터가 [EndedChatRoomLoader]로 종료 결과를
  * 조립해 [MemberReviewService.createReviews]에 넘긴다.
  *
  * **채팅 종료와 평가 생성을 한 트랜잭션으로 묶지 않는다.** 평가 생성 실패가 채팅 종료를 되돌리면
  * 사용자가 나가기를 눌렀는데 실패하기 때문이다. 대신 계약을 at-least-once 로 두고, 놓친 방은
  * [openMissing]이 줍는다. `createReviews`가 멱등이라 중복 생성은 일어나지 않는다.
  *
- * **그룹은 평가보다 재매칭 쌍을 먼저 만든다.** 그룹 평가는 재매칭 의사를 필수로 받고 `RematchSubmitter`가
- * 쌍을 찾지 못하면 제출을 거부하므로, 순서가 뒤집히면 사용자가 평가를 다 채우고 제출에서 막힌다.
+ * **그룹은 평가보다 재매칭 쌍을 먼저 만든다**(그 순서여야 하는 이유는 [RematchPairCreator]).
  * 중간에 실패해 "쌍만 있고 평가 없음"으로 남는 것은 무해하다 — 쌍은 평가 없이는 쓰이지 않고,
  * 다음 복구 주기가 평가를 열면서 이미 있는 쌍을 건너뛴다.
  */
@@ -32,12 +25,9 @@ import org.springframework.stereotype.Component
 class EndedChatReviewOpener(
     private val memberReviewService: MemberReviewService,
     private val rematchPairCreator: RematchPairCreator,
+    private val endedChatRoomLoader: EndedChatRoomLoader,
     private val memberReviewRepository: MemberReviewRepository,
     private val chatRoomRepository: ChatRoomRepository,
-    private val chatRoomMemberRepository: ChatRoomMemberRepository,
-    private val personalMatchRepository: PersonalMatchRepository,
-    private val groupMatchRepository: GroupMatchRepository,
-    private val quizSetRepository: QuizSetRepository,
 ) {
     /**
      * 방금 끝난 방들의 평가를 곧바로 연다. 종료 응답이 나가기 전에 평가가 열려 있어야
@@ -51,7 +41,7 @@ class EndedChatReviewOpener(
             return
         }
         // 방마다 격리되므로 여기서 전체를 감쌀 필요가 없다. 조회 자체가 실패하는 경우만 남는다.
-        runCatchingExceptions { openEach(chatRoomRepository.findAllById(endedRoomIds)) }
+        runCatchingExceptions { openRooms(chatRoomRepository.findAllById(endedRoomIds)) }
             .onFailure { logger.warn(it) { "종료 직후 평가 열기 실패 — 누락 복구에 맡긴다: roomIds=$endedRoomIds" } }
     }
 
@@ -60,6 +50,12 @@ class EndedChatReviewOpener(
      *
      * [openFor]가 실패했거나 그 사이 앱이 죽어 평가가 안 열린 방을 줍는다. 종료 시각 하한은 두지 않는다 —
      * 얼마나 오래 밀렸든 복구되어야 하기 때문이다(근거는 리포지토리 메서드 KDoc).
+     *
+     * [openFor]와 달리 자신의 조회 실패는 흡수하지 않는다. 부르는 쪽이 크론이라 예외가 올라가도
+     * 그 주기만 건너뛰고 다음 주기에 다시 시도되는 반면, [openFor]는 사용자 종료 요청 경로에 있어
+     * 예외가 종료 실패로 보이기 때문이다.
+     *
+     * @return 평가가 실제로 열린 방 수
      */
     fun openMissing(): Int {
         val roomIds = memberReviewRepository.findEndedChatRoomIdsWithoutReview(RECOVERY_BATCH_SIZE)
@@ -68,7 +64,7 @@ class EndedChatReviewOpener(
         }
 
         val rooms = chatRoomRepository.findAllById(roomIds)
-        val opened = openEach(rooms)
+        val opened = openRooms(rooms)
         logger.info { "평가 누락 복구: 대상 ${rooms.size}건 중 ${opened}건 열림" }
         return opened
     }
@@ -81,110 +77,20 @@ class EndedChatReviewOpener(
      * 뒤쪽 한 건 때문에 앞서 성공한 것까지 폐기되고, anti-join 이 그 방을 매 주기 다시 집어오므로
      * **독이 든 방 하나가 복구를 영구히 막는다**(예: 참여자 0명이면 `createReviews`가 예외를 던진다).
      *
-     * 그룹은 건너뛴다 — 멤버십 동결·재매칭 pair 가 얽혀 별도 트랙(`I1G`)이다.
-     *
      * @return 평가가 실제로 열린 방 수
      */
-    private fun openEach(rooms: List<ChatRoom>): Int =
-        loadEndedChatRooms(rooms)
+    private fun openRooms(rooms: List<ChatRoom>): Int =
+        endedChatRoomLoader.load(rooms)
             .count { endedChatRoom ->
-                runCatchingExceptions { openOne(endedChatRoom) }
+                runCatchingExceptions { openRoom(endedChatRoom) }
                     .onFailure { logger.warn(it) { "평가 열기 실패 — 다음 복구 주기로 넘긴다: roomId=${endedChatRoom.chatRoomId}" } }
                     .isSuccess
             }
 
     /** 그룹이면 재매칭 쌍을 먼저 만든 뒤 평가를 연다. 1:1 은 쌍 생성이 no-op 이다. */
-    private fun openOne(endedChatRoom: EndedChatRoom) {
+    private fun openRoom(endedChatRoom: EndedChatRoom) {
         rematchPairCreator.createPairsFor(endedChatRoom)
         memberReviewService.createReviews(endedChatRoom)
-    }
-
-    /**
-     * 방마다 평가 입력 계약([EndedChatRoom])을 만든다. `chat_room`에 없는 값을 원본 매칭에서 읽어 채우므로
-     * 순수 변환이 아니다 — `quizSetId`는 `PersonalMatch`, `weekStartedOn`은 그 퀴즈셋에 있다.
-     *
-     * 방 수만큼 조회하지 않도록 필요한 것을 먼저 일괄로 읽은 뒤 방마다 짝지운다.
-     * 값이 빠진 방은 [toEndedChatRoom]이 건너뛰므로, 돌려주는 목록이 입력보다 짧을 수 있다.
-     */
-    private fun loadEndedChatRooms(rooms: List<ChatRoom>): List<EndedChatRoom> {
-        if (rooms.isEmpty()) {
-            return emptyList()
-        }
-
-        val quizSetIdByRoomId = findQuizSetIdByRoomId(rooms)
-        val weekStartedOnByQuizSetId = findWeekStartedOnByQuizSetId(quizSetIdByRoomId.values)
-        val participantIdsByRoomId = findParticipantIdsByRoomId(rooms)
-
-        return rooms.mapNotNull { room ->
-            toEndedChatRoom(
-                room = room,
-                quizSetId = quizSetIdByRoomId[room.id],
-                weekStartedOnByQuizSetId = weekStartedOnByQuizSetId,
-                participantIds = participantIdsByRoomId[room.id].orEmpty(),
-            )
-        }
-    }
-
-    /**
-     * 방마다 원본 매칭을 타고 `quizSetId`를 찾는다. 원본이 유형별로 다른 테이블이라
-     * (`PERSONAL`은 `personal_match`, `GROUP`은 `group_match`) 여기서 흡수하고,
-     * 이후 조립은 유형을 신경 쓰지 않는다. 원본이 없는 방은 map 에 담기지 않는다.
-     */
-    private fun findQuizSetIdByRoomId(rooms: List<ChatRoom>): Map<Long, Long> {
-        val (personalRooms, groupRooms) = rooms.partition { it.sourceType == ChatRoomType.PERSONAL }
-        val personalQuizSetIdByMatchId = personalMatchRepository.findAllById(personalRooms.map { it.sourceId })
-            .associate { it.id to it.quizSetId }
-        val groupQuizSetIdByMatchId = groupMatchRepository.findAllById(groupRooms.map { it.sourceId })
-            .associate { it.id to it.quizSetId }
-
-        return rooms.mapNotNull { room ->
-            val quizSetId = when (room.sourceType) {
-                ChatRoomType.PERSONAL -> personalQuizSetIdByMatchId[room.sourceId]
-                ChatRoomType.GROUP -> groupQuizSetIdByMatchId[room.sourceId]
-            }
-            quizSetId?.let { room.id to it }
-        }.toMap()
-    }
-
-    private fun findWeekStartedOnByQuizSetId(quizSetIds: Collection<Long>): Map<Long, LocalDate> =
-        quizSetRepository.findAllById(quizSetIds).associate { it.id to it.weekStartedOn }
-
-    private fun findParticipantIdsByRoomId(rooms: List<ChatRoom>): Map<Long, List<Long>> =
-        chatRoomMemberRepository.findByRoomIdIn(rooms.map { it.id }).groupBy({ it.roomId }, { it.memberId })
-
-    /**
-     * 조립에 필요한 값이 하나라도 없으면 그 방은 건너뛴다.
-     *
-     * 원본이 사라졌거나(탈퇴 hard delete 등) 종료 시각이 없는 방이 그렇다. 여기서 터뜨리면
-     * 같은 배치의 정상 방들까지 막히므로, 건너뛰고 로그로만 남겨 다음 복구 주기에 다시 시도한다.
-     */
-    private fun toEndedChatRoom(
-        room: ChatRoom,
-        quizSetId: Long?,
-        weekStartedOnByQuizSetId: Map<Long, LocalDate>,
-        participantIds: List<Long>,
-    ): EndedChatRoom? {
-        val weekStartedOn = quizSetId?.let { weekStartedOnByQuizSetId[it] }
-        val endedAt = room.endedAt
-
-        if (quizSetId == null || weekStartedOn == null || endedAt == null) {
-            logger.warn {
-                "평가를 열 수 없어 건너뜀: roomId=${room.id}, sourceType=${room.sourceType}, " +
-                    "sourceId=${room.sourceId}, quizSetId=$quizSetId, " +
-                    "weekStartedOn=${weekStartedOn != null}, endedAt=$endedAt"
-            }
-            return null
-        }
-
-        return EndedChatRoom(
-            chatRoomId = room.id,
-            matchType = room.sourceType,
-            matchId = room.sourceId,
-            quizSetId = quizSetId,
-            weekStartedOn = weekStartedOn,
-            participantIds = participantIds,
-            endedAt = endedAt,
-        )
     }
 
     /**
@@ -202,6 +108,5 @@ class EndedChatReviewOpener(
 
         /** 한 번의 복구가 떠안을 최대 방 수. 장애 후 밀린 물량이 한 호출을 오래 잡지 않게 끊는다. */
         private const val RECOVERY_BATCH_SIZE = 100
-
     }
 }
